refactor: Change `/component/settings` to `/component/room-settings`
- Making the component's purpose more explicit as room-specific settings
